refactor: move app-level provides into separate plugin
closes #330
